# semantic-modeling-assistant-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/legal-acts/{year}/{number}/{date}/class-suggestions-top-k-extraction-jobs": {
      "post": {
        "summary": "Start a new top-K class suggestions extraction job",
        "description": "Starts a new job to extract the top-K class suggestions for specific paragraphs (structural elements) of the given legal act. After the job is started, it will run asynchronously and can be checked for status using the job ID.",
        "operationId": "start_class_suggestions_top_k_extraction_job_legal_acts__year___number___date__class_suggestions_top_k_extraction_jobs_post",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Official number of the legal act",
              "title": "Number"
            },
            "description": "Official number of the legal act"
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Year of the legal act",
              "title": "Year"
            },
            "description": "Year of the legal act"
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date identifying the version of the legal act (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Date identifying the version of the legal act (YYYY-MM-DD)"
          },
          {
            "name": "user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User-Id"
            }
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartClassSuggestionsTopKExtractionJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartClassSuggestionsJobResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "You have reached the maximum usage available to you. Please try again later."
                    }
                  },
                  "required": [
                    "detail"
                  ],
                  "title": "UnauthorizedError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/legal-acts/{year}/{number}/{date}/class-suggestions-jobs/{job_id}": {
      "get": {
        "summary": "Get job status",
        "description": "Get the status of a class suggestions job by job ID received when starting the job.The status contains the status information about the job and the list of new class suggestions extracted since the job was started. After the job is completed, no more suggestions will be added to this list.",
        "operationId": "get_class_suggestion_job_status_legal_acts__year___number___date__class_suggestions_jobs__job_id__get",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Official number of the legal act",
              "title": "Number"
            },
            "description": "Official number of the legal act"
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Year of the legal act",
              "title": "Year"
            },
            "description": "Year of the legal act"
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date identifying the version of the legal act (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Date identifying the version of the legal act (YYYY-MM-DD)"
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Unique identifier of the job",
              "title": "Job Id"
            },
            "description": "Unique identifier of the job"
          },
          {
            "name": "user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User-Id"
            }
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassSuggestionsJobStatusResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "You have reached the maximum usage available to you. Please try again later."
                    }
                  },
                  "required": [
                    "detail"
                  ],
                  "title": "UnauthorizedError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/legal-acts/{year}/{number}/{date}/property-suggestions-jobs/{job_id}": {
      "get": {
        "summary": "Get property (attribute and relationship) suggestion job status",
        "description": "Get the status of a property (attribute and relationship) suggestions job by job ID received when starting the job. The status contains the status information about the job and the list of new property (attribute and relationship) suggestions extracted since the job was started. After the job is completed, no more suggestions will be added to this list.",
        "operationId": "get_property_suggestion_job_status_legal_acts__year___number___date__property_suggestions_jobs__job_id__get",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Official number of the legal act",
              "title": "Number"
            },
            "description": "Official number of the legal act"
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Year of the legal act",
              "title": "Year"
            },
            "description": "Year of the legal act"
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date identifying the version of the legal act (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Date identifying the version of the legal act (YYYY-MM-DD)"
          },
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Job ID",
              "title": "Job Id"
            },
            "description": "Job ID"
          },
          {
            "name": "user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User-Id"
            }
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySuggestionsJobStatusResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "You have reached the maximum usage available to you. Please try again later."
                    }
                  },
                  "required": [
                    "detail"
                  ],
                  "title": "UnauthorizedError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/legal-acts/{year}/{number}/{date}/property-suggestions-top-k-extraction-jobs": {
      "post": {
        "summary": "Start a top-K attribute suggestions extraction job",
        "description": "Starts a job to extract the top-K attribute suggestions for specific structural elements of a legal act and for a specific class. After the job is started, it will run asynchronously and can be checked for status using the job ID.",
        "operationId": "start_property_suggestions_top_k_extraction_job_legal_acts__year___number___date__property_suggestions_top_k_extraction_jobs_post",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Official number of the legal act",
              "title": "Number"
            },
            "description": "Official number of the legal act"
          },
          {
            "name": "year",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Year of the legal act",
              "title": "Year"
            },
            "description": "Year of the legal act"
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Date identifying the version of the legal act (YYYY-MM-DD)",
              "title": "Date"
            },
            "description": "Date identifying the version of the legal act (YYYY-MM-DD)"
          },
          {
            "name": "user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User-Id"
            }
          },
          {
            "name": "password",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPropertySuggestionsTopKExtractionJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPropertySuggestionsJobResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "example": "You have reached the maximum usage available to you. Please try again later."
                    }
                  },
                  "required": [
                    "detail"
                  ],
                  "title": "UnauthorizedError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassSuggestionsJobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id",
            "description": "Unique identifier of the job"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current status of the job ('in_progress', 'completed', 'failed'"
          },
          "new_suggestions": {
            "items": {
              "$ref": "#/components/schemas/ClassSuggestion"
            },
            "type": "array",
            "title": "New Suggestions",
            "description": "List of extracted class suggestions since the start of the job. After 'completed' status is set, this list contains all suggestions generated by the job. No more suggestions will be added after the job is completed.",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "ClassSuggestionsJobStatusResponse",
        "description": "Represents the status of a class suggestions extraction job with the list of already extracted class suggestions."
      },
      "ConceptualModel": {
        "properties": {
          "classes": {
            "items": {
              "$ref": "#/components/schemas/Class"
            },
            "type": "array",
            "title": "Classes",
            "description": "List of classes in the conceptual model"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/Relationship"
            },
            "type": "array",
            "title": "Relationships",
            "description": "List of relationships in the conceptual model"
          }
        },
        "type": "object",
        "title": "ConceptualModel",
        "description": "Represents a conceptual model consisting of classes and relationships."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PropertySuggestionsJobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id",
            "description": "Unique identifier of the job"
          },
          "selected_class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Class Id",
            "description": "The ID of the class for which property suggestions are being generated"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current status of the job ('in_progress', 'completed', 'failed')"
          },
          "new_relationship_suggestions": {
            "items": {
              "$ref": "#/components/schemas/RelationshipSuggestion"
            },
            "type": "array",
            "title": "New Relationship Suggestions",
            "description": "List of extracted relationship suggestions since the start of the job. After 'completed' status is set, this list contains all suggestions generated by the job. No more suggestions will be added after the job is completed.",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "selected_class_id",
          "status"
        ],
        "title": "PropertySuggestionsJobStatusResponse"
      },
...
      "StartClassSuggestionsJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id",
            "description": "Unique identifier for the started job"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Current status of the job ('in_progress', 'completed', 'failed')"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "StartClassSuggestionsJobResponse",
        "description": "Represents a response to a request for starting a class suggestions extraction job."
      },
      "StartClassSuggestionsTopKExtractionJobRequest": {
        "properties": {
          "k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "K",
            "description": "Maximal number of top class suggestions to extract in this job (default: 10)",
            "default": 10
          },
          "structural_element_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structural Element Ids",
            "description": "List of paragraph numbers (structural element IDs) to process by the class extraction job. If not provided, the whole legal act will be processed. However, take into account that processing the whole legal act can exceed the tokens-per-minute limit given by the LLM API provider which would lead to job rejection."
          },
          "context_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text",
            "description": "Additional context text specified by the user to focus suggestion extraction. Be careful in using this as it is not guaranteed to be used in the extraction process. Also take into account that it can break the extraction process if specified maliciously. The job can also be rejected completely if the context is found harmful."
          },
          "known_conceptual_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConceptualModel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Existing conceptual model to consider during the extraction job. The extraction job will try to generate new suggestions that are semantically close to the existing model and it will try to avoid generating suggestions that are already present in the existing model. If not provided, the job will generate suggestions without considering any existing model."
          }
        },
        "type": "object",
        "title": "StartClassSuggestionsTopKExtractionJobRequest",
        "description": "Represents a request to start a job for extracting top-K class suggestions from the given set of paragraphs (structural elements) of a legal act."
      },
      "StartPropertySuggestionsJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id",
            "description": "Unique identifier for the started job"
          },
          "selected_class_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Class Id",
            "description": "The ID of the class for which property suggestions are to be generated"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Current status of the job ('in_progress', 'completed', 'failed')"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "selected_class_id"
        ],
        "title": "StartPropertySuggestionsJobResponse",
        "description": "Represents a response to a request for starting a property (attribute, relationship) suggestions extraction job."
      },
      "StartPropertySuggestionsTopKExtractionJobRequest": {
        "properties": {
          "k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "K"
          },
          "selected_class_id": {
            "type": "string",
            "title": "Selected Class Id",
            "description": "The ID of the class for which property suggestions are to be generated. This includes attributes owned by the class and relationships mediating the class with another class. If the other class does not exist in the known conceptual model, it will extracted as a new class suggestion but only with name, its definition and explanation will not be generated."
          },
          "structural_element_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structural Element Ids",
            "description": "The IDs of the structural elements (paragraphs) from which to extract property suggestions. If not provided, the whole legal text will be considered. However, take into account that in that case, the limits given by the LLM API provider, e.g. context window size or transactions-per-minute, can be exceeded resulting in the job failure."
          },
          "context_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Text",
            "description": "Additional context text specified by the user to focus suggestion extraction. Be careful in using this as it is not guaranteed to be used in the extraction process. Also take into account that it can break the extraction process if specified maliciously. The job can also be rejected completely if the context is found harmful."
          },
          "known_conceptual_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConceptualModel"
              },
              {
                "type": "null"
              }
            ],
            "description": "Existing conceptual model to consider during the extraction job. The extraction job will try to generate new suggestions that are semantically close to the existing model and it will try to avoid generating suggestions that are already present in the existing model. If not provided, the job will generate suggestions without considering any existing model."
          }
        },
        "type": "object",
        "required": [
          "k",
          "selected_class_id"
        ],
        "title": "StartPropertySuggestionsTopKExtractionJobRequest",
        "description": "Represents a request to start a job for extracting top-K property suggestions (i.e. mix of attributes and relationships) from the given set of paragraphs (structural elements) of a legal act for the given class."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}